♻️ refactor: split trace hook by domain (llm, tool, memory)
Each file has the OnXxx methods with both slog + OTel logic together:
- hook.go: types, init, constructor, OTel provider, lifecycle
- llm.go: OnPreLLMCall, OnPostLLMCall
- tool.go: OnPreToolCall, OnPostToolCall, summarizeArgs
- memory.go: OnPostMemoryCall
- config.go: env var config
